ASoC: intel/sdw_utils: move soundwire machine driver helper functions
Move below Intel SoundWire machine driver helper functions to
soc_sdw_utils.c file so that it can be used by other platform machine
driver.
- is_unique_device()
- get_codec_name()
- get_codec_dai_by_name()
Rename the functions by adding 'asoc_sdw' tag as a prefix.
