+++ sound/soc/intel/boards/sof_sdw_cs42l42.c
@@ -59,7 +59,7 @@ int cs42l42_rtd_init(struct snd_soc_pcm_runtime *rtd)
 	struct snd_soc_jack *jack;
 	int ret;
 
-	codec_dai = get_codec_dai_by_name(rtd, jack_codecs, ARRAY_SIZE(jack_codecs));
+	codec_dai = asoc_sdw_get_codec_dai_by_name(rtd, jack_codecs, ARRAY_SIZE(jack_codecs));
 	if (!codec_dai)
 		return -EINVAL;
